pyproject.toml
[tool.poetry]
name = "affex_app"
version = "0.1.0"
description = "A social media backend application"
authors = ["romeo <romeomeshack34@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0.6"
djangorestframework = "3.13.1"
PyJWT = "2.4.0"
gunicorn = "20.1.0"
djangorestframework-simplejwt = "5.2.0"
drf-spectacular = "0.22.1"
python-decouple = "3.6"
dj-database-url = "^1.0.0"
s3transfer = "0.6.0"
django-cors-headers = "3.13.0"
django-extensions = "3.2.0"
django-storages = "1.12.3"
python-dateutil = "2.8.2"
PyYAML = "6.0"
redis = "4.3.4"
psycopg2-binary = "^2.9.3"
boto3 = "1.24.38"
colorlog = "6.6.0"
celery = "5.2.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
